requirements.txt
aiogram>=3.4.1
openai>=1.52.0
python-dotenv>=1.0.0
httpx<0.28.0
asyncpg>=0.29.0
apscheduler>=3.10.4
